feat(salvageunion-reference): move Tier 2 rules modules into the package (ADR-006)
Migrates the remaining "small outcome-shape types" rules modules,
completing the design's Tier 2:

- takeDamage.ts — moved wholesale. Its outcome/result types
  (CriticalDamageOutcome/Result, CriticalInjuryOutcome/Result) were
  actually Zod-inferred imports from ITUN's src/lib/schemas/{mech,pilot}.ts
  (a discrepancy from the original per-module survey, which had assumed
  zero-schema-import for this tier) — generalized into the same kind of
  small structural outcome-shape type the design already prescribed for
  derivedStats.ts. ITUN's Zod types satisfy them structurally.
- heatCheck.ts — moved ONLY the pure effect computation (clampHeat,
  reactorOverloadOutcome, performHeatCheck, performPush). ITUN's
  heatCheck.ts stays as a hybrid shim: it re-exports the pure functions
  from the package, but keeps `defaultRoll` (@randsum/roller binding) and
  `heatCheckPatch` (Partial<Mech> patch assembly) local, per the design.
  Its HeatCheckResult/ReactorOverloadOutcome/HeatCheckEffect types also
  turned out to be Zod-derived (same discrepancy as takeDamage) and are
  now structural types in the package.
- coreMechanic.ts — moved wholesale. It's nominally Tier 1 in the design
  (zero ITUN-schema imports) but actually imports `Roll`/`HeatCheckEffect`
  from ITUN's heatCheck.ts (a Tier 2 module) — so it was moved in this
  commit, after heatCheck's pure split landed, resolving that dependency
  edge in package-internal order instead of forcing a Tier-1-before-Tier-2
  sequencing that doesn't reflect the actual dependency graph.
- derivedStats.ts — generalized its `Pick<Mech/Pilot/Crawler, ...>`
  parameter types into local plain structural types (small outcome/result
  record shapes, not full persisted records) that ITUN's real Zod types
  satisfy structurally. `crawlerMaxSP` needed
  `parseCrawlerTechLevel` (a 4-line slug-digit parser from ITUN's
  src/lib/crawlerLevel.ts) — duplicated locally rather than imported,
  since crawlerLevel.ts also owns `resolveEffectiveCrawlerLevel`, which
  depends on ITUN's Pilot/Crawler Zod types and is out of this migration's
  scope; importing the file would have pulled that dependency in too.
- mediatorTables.ts — moved wholesale; its MediatorRollResult/
  MediatorTableId types were Zod-inferred from ITUN's
  src/lib/schemas/encounterNpc.ts (same discrepancy pattern) and are now
  package structural types.

ITUN's local module files are thin re-export shims (heatCheck.ts is a
hybrid: pure parts re-exported, defaultRoll/heatCheckPatch stay local).
Consumers import by submodule path, not a barrel, so this needed zero
consumer-file changes.

Verified: full-monorepo typecheck clean (0 errors), all test suites
green (package: 198 rules tests; ITUN: 1116 tests, the ~108 that moved
now run in the package instead), lint clean, build:package produces no
schema drift, validate:all passes.
