+++ packages/salvageunion-reference/lib/rules/heatCheck.test.ts
@@ -8,8 +8,8 @@
 
 import { describe, expect, test } from 'bun:test'
 
-import { clampHeat, performHeatCheck, performPush, reactorOverloadOutcome } from '../heatCheck'
-import type { Roll } from '../heatCheck'
+import { clampHeat, performHeatCheck, performPush, reactorOverloadOutcome } from './heatCheck.js'
+import type { Roll } from './types.js'
 
 /** Returns a Roll that yields the given values in order, ignoring `sides`. */
 function seqRoll(...values: number[]): Roll {
